move configs to their own file
