--- mminf/model/bagel/utils.py
@@ -1,36 +1,0 @@
-def add_special_tokens(tokenizer):
-    all_special_tokens = []
-    for k, v in tokenizer.special_tokens_map.items():
-        if isinstance(v, str):
-            all_special_tokens.append(v)
-        elif isinstance(v, list):
-            all_special_tokens += v
-
-    new_tokens = []
-
-    if '<|im_start|>' not in all_special_tokens:
-        new_tokens.append('<|im_start|>')
-
-    if '<|im_end|>' not in all_special_tokens:
-        new_tokens.append('<|im_end|>')
-
-    if '<|vision_start|>' not in all_special_tokens:
-        new_tokens.append('<|vision_start|>')
-
-    if '<|vision_end|>' not in all_special_tokens:
-        new_tokens.append('<|vision_end|>')
-
-    num_new_tokens = tokenizer.add_tokens(new_tokens)
-    bos_token_id = tokenizer.convert_tokens_to_ids('<|im_start|>')
-    eos_token_id = tokenizer.convert_tokens_to_ids('<|im_end|>')
-    start_of_image = tokenizer.convert_tokens_to_ids('<|vision_start|>')
-    end_of_image = tokenizer.convert_tokens_to_ids('<|vision_end|>')
-
-    new_token_ids = dict(
-        bos_token_id=bos_token_id, 
-        eos_token_id=eos_token_id, 
-        start_of_image=start_of_image, 
-        end_of_image=end_of_image, 
-    )
-
-    return tokenizer, new_token_ids, num_new_tokens
